try to establish connection
